api/pyproject.toml
[tool.poetry]
name = "grid-api"
version = "0.1.0"
description = ""
authors = ["AntonZakhlebayeu <antonzakhlebayeu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.116.1"
sqlalchemy = "^2.0.41"
uvicorn = "^0.35.0"
psycopg2-binary = "^2.9.10"
python-dotenv = "^1.1.1"
pydantic = "^2.11.7"
pydantic-settings = "^2.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
